pyproject.toml
[project]
name = "01-agent"
version = "0.1.0"
description = "My first attempt at building an agent. Use case: project manager on Trello"
authors = [
    {name = "ladams537",email = "louistadams3@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "pydantic-ai[google] (>=1.27.0,<2.0.0)",
    "devtools (>=0.12.2,<0.13.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-genai (>=1.53.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
